# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqAdmin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IAdmin": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginAdmin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrder": {
				"properties": {
					"bookName": {
						"type": "string"
					},
					"issuanceDate": {
						"type": "string",
						"format": "date-time"
					},
					"returnDate": {
						"type": "string",
						"format": "date-time"
					},
					"fees": {
						"type": "number",
						"format": "double"
					},
					"fine": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"bookName",
					"issuanceDate",
					"returnDate",
					"fees",
					"fine"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResCustomer": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"order": {
						"items": {
							"$ref": "#/components/schemas/IOrder"
						},
						"type": "array"
					},
					"totalFees": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"order",
					"totalFees",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqCustomer": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"order": {
						"items": {
							"$ref": "#/components/schemas/IOrder"
						},
						"type": "array"
					},
					"totalFees": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"order",
					"totalFees"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCustomer": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqCustomer": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"order": {
						"items": {
							"$ref": "#/components/schemas/IOrder"
						},
						"type": "array"
					},
					"totalFees": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"password",
					"order",
					"totalFees"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICustomer": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteCustomer": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResBook": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"desc",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqBook": {
				"properties": {
					"name": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"desc"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetBook": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqBook": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"desc"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBook": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteBook": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "library-management",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Admin/saveAdmin": {
			"post": {
				"operationId": "SaveAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqAdmin"
							}
						}
					}
				}
			}
		},
		"/Admin/updateAdmin": {
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqAdmin"
							}
						}
					}
				}
			}
		},
		"/Admin/getAdmin": {
			"post": {
				"operationId": "GetAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetAdmin"
							}
						}
					}
				}
			}
		},
		"/Admin/getAdminList": {
			"get": {
				"operationId": "GetAdminList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResAdmin"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Admin/deleteAdmin": {
			"delete": {
				"operationId": "DeleteAdmin",
				"responses": {
					"200": {
						"description": "Admin Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IAdmin"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAdmin"
							}
						}
					}
				}
			}
		},
		"/Admin/loginAdmin": {
			"post": {
				"operationId": "LoginAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginAdmin"
							}
						}
					}
				}
			}
		},
		"/customer/saveCustomer": {
			"post": {
				"operationId": "SaveCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResCustomer"
								}
							}
						}
					}
				},
				"summary": "Summary for saving customer",
				"tags": [
					"Customer Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqCustomer"
							}
						}
					}
				}
			}
		},
		"/customer/getCustomer": {
			"post": {
				"operationId": "GetCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResCustomer"
								}
							}
						}
					}
				},
				"tags": [
					"Customer Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetCustomer"
							}
						}
					}
				}
			}
		},
		"/customer/getCustomerList": {
			"get": {
				"operationId": "GetCustomerList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResCustomer"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Customer Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/customer/updateCustomer": {
			"put": {
				"operationId": "UpdateCustomer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResCustomer"
								}
							}
						}
					}
				},
				"tags": [
					"Customer Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqCustomer"
							}
						}
					}
				}
			}
		},
		"/customer/deleteCustomer": {
			"delete": {
				"operationId": "DeleteCustomer",
				"responses": {
					"200": {
						"description": "Customer Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ICustomer"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Customer Routes"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteCustomer"
							}
						}
					}
				}
			}
		},
		"/book/saveBook": {
			"post": {
				"operationId": "SaveBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBook"
								}
							}
						}
					}
				},
				"tags": [
					"Book Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqBook"
							}
						}
					}
				}
			}
		},
		"/book/getBook": {
			"post": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBook"
								}
							}
						}
					}
				},
				"tags": [
					"Book Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetBook"
							}
						}
					}
				}
			}
		},
		"/book/getBookList": {
			"get": {
				"operationId": "GetBookList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResBook"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Book Routes"
				],
				"security": [],
				"parameters": []
			}
		},
		"/book/updateBook": {
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResBook"
								}
							}
						}
					}
				},
				"tags": [
					"Book Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqBook"
							}
						}
					}
				}
			}
		},
		"/book/deleteBook": {
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"200": {
						"description": "Book deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IBook"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Book Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteBook"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}